don't emit all accessors for all messages
Allow specifying the "type" of the QMI package, type can be one of
"client", "server" or "agnostic" with "agnostic" being the default for
backwards compability.

Client applications don't need to emit getter methods on QMI requests,
and likewise don't need setters on responses. They don't need to
_alloc() requests, etc. This change avoids emitting these extra
accessors, drastically reducing the size of the emitted code.