WIP: accessor: superseed kernel-style generator
Generate a struct for each QMI message and add a getall method to call
each getter and populate the struct. All pointer types must be free'd
the struct doesn't share any memory with the response buffer, so the
response message itself can be free'd without invalidating the struct.

TODO: Generate setter code as well!
TODO: Generate _free() method (also, this needs to be done for
deserialised structs for sure oops)